extracted Printer module
