+++ Gemfile
@@ -2,6 +2,12 @@
 
 source 'https://rubygems.org'
 
+ruby "2.7.1"
+
+gem 'rake'
+
+group :test do
 gem 'rspec'
 gem 'rubocop', require: false
-gem 'simplecov', require: false, group: :test
+gem 'simplecov', require: false
+end
