+++ spec/transaction_spec.rb
@@ -1,0 +1,8 @@
+# # frozen_string_literal: true
+
+# describe Transaction do
+  
+#   it 'contains the date, credit, debit and balance'
+#     expect(subject.)
+#   end
+# end
